+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## unreleased
+- Add `AcquirerReferenceNumber` to `Transaction`
+
 ## 3.0.0 
 * Add `RequestTimeoutError` and `GatewayTimeoutError` exceptions
 * Remove several no-longer-used dependencies (underscore, semver and readable-stream)